% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_function.R
\name{add_deprecated_call}
\alias{add_deprecated_call}
\title{Add `.Deprecated()` to Text}
\usage{
add_deprecated_call(text, new = NULL)
}
\arguments{
\item{text}{Character vector containing function declarations.}

\item{new}{(Optional) Argument passed to the new call to `.Deprecated()`.}

\item{package}{PARAM_DESCRIPTION, Default: NULL}
}
\description{
`.Deprecated()` is added after each function declaration in the text. Note that it does not check for an existing `.Deprecated()`.
}
\examples{
library(tidyverse)

# Adds after each function declaration
add_deprecated_call("sample_function1 <-\nfunction(a,b,c){}\n\n#' Roxygen2 Documentation\n#' Roxygen2 Lines\n\nsample_function2 <- \nfunction(x,y,z){}") \%>\% cat()

# Does not skip function declarations that already makes a call to the .Deprecated function
add_deprecated_call("sample_function1 <-\nfunction(a,b,c){\n.Deprecated()\n}\n\n#' Roxygen2 Documentation\n#' Roxygen2 Lines\n\nsample_function2 <- \nfunction(x,y,z){}") \%>\% cat()

# Text can be a vector of length greater than 1
add_deprecated_call(c("sample_function1 <-\nfunction(a,b,c){\n.Deprecated()\n}", "#' Roxygen2 Documentation\n#' Roxygen2 Lines\n\nsample_function2 <- \nfunction(x,y,z){}")) \%>\% cat()
}
\seealso{
\code{\link[stringr]{str_replace}}
}
\concept{modify roxygen2 documentation functions}
