% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_ext.R
\name{append_ext}
\alias{append_ext}
\title{Append an Extension to a Filename}
\usage{
append_ext(filename, ext)
}
\arguments{
\item{filename}{Extension-less filename. For example, "IngestionFile".}

\item{ext}{Extension to append withou the period. For example, "csv".}
}
\description{
Append an extension to an extension-less filename
}
