% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_path.R
\name{append_path}
\alias{append_path}
\title{Append a Path to a Filename}
\usage{
append_path(filename, path)
}
\arguments{
\item{filename}{Path-less filename. For example, "IngestionFile.xlsx".}

\item{path}{Path to append without the terminal forward-slash. For example, "~/GitHub".}
}
\description{
Append an path to an path-less filename
}
