% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_object.R
\name{cache_object}
\alias{cache_object}
\title{Cache an object for retrieval later}
\usage{
cache_object(object)
}
\arguments{
\item{object}{object to cache.}
}
\description{
This function uses the R.cache package, but creates a key automatically based on the present R script and the object name. The directory for the cache also by default is in the set path with a subdirectory that matches the name of the working directory.
}
