% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-wd.R
\name{change_wd}
\alias{change_wd}
\title{Set a new wd while saving the previous one}
\usage{
change_wd(new_wd, save = TRUE)
}
\description{
On the condition that the new_wd is not the same as the current wd, this function sets a new working directory while saving the previous wd as a `wd0` object. If the condition is not met, no new wd is set and a `wd0` object will not be returned to the global environment.
}
