% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_line.R
\name{cl_open_ff}
\alias{cl_open_ff}
\title{Open a File in a Specific Application}
\usage{
cl_open_ff(app)
}
\arguments{
\item{app}{Path to the Application}
}
\description{
Open a File in a Specific Application
}
