% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_object.R
\name{copy_object}
\alias{copy_object}
\title{Copy an R object to another R object of a different name}
\usage{
copy_object(new_object_name, object_to_copy_name)
}
\arguments{
\item{new_object_name}{string of new object name}

\item{object_to_copy_name}{string of object name to b copied}
}
\description{
Copy an R object to another R object of a different name
}
