% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdir_if_not_exist.R
\name{create_dir_if_not_exist}
\alias{create_dir_if_not_exist}
\title{Creates folder in working dir if doesn't exist}
\usage{
create_dir_if_not_exist(folder_name)
}
\arguments{
\item{folder_name}{name of the directory}
}
\description{
Creates folder in working dir if doesn't exist
}
