% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cut_cache_key}
\alias{cut_cache_key}
\title{Create the key for caching}
\usage{
cut_cache_key(object)
}
\description{
The list object that serves as a key for the caching functions of the R.cache package. This function differs from use_key in that this one is used to first save a virgin object while the use_key is used to retrieve an object.
}
