% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate_all_csv}
\alias{deprecate_all_csv}
\title{Deprecate all csvs in wd}
\usage{
deprecate_all_csv()
}
\description{
A deprecated/ dir is created if it does not already exist and all files with the pattern "csv$" are moved to this folder.
}
