% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate_all_excel}
\alias{deprecate_all_excel}
\title{Deprecate all Excels in wd}
\usage{
deprecate_all_excel()
}
\description{
A deprecated/ dir is created if it does not already exist and all files with the pattern "xlsx$" are moved to this folder.
}
