% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate_all_txt}
\alias{deprecate_all_txt}
\title{Deprecate all txt files in wd}
\usage{
deprecate_all_txt()
}
\description{
A deprecated/ dir is created if it does not already exist and all files with the pattern "txt$" are moved to this folder.
}
