% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_mgmt.R
\name{dir.create_path}
\alias{dir.create_path}
\title{Create the Path to a Directory}
\usage{
dir.create_path(dir)
}
\description{
Create a directory along with the intermediate directories if any do not already exist. A console message is returned indicating whether the directory already existed or if it was created by the function call.
}
\examples{

# Create a Directory Path
dir.create_path("~/Desktop/test")
dir.create_path("~/Desktop/test2")
dir.create_path("~/Desktop/test2/test3")
dir.create_path("~/Desktop/test2/test3/test4/test5")


# Remove Directory
remove.dir("~/Desktop/test2/test3/test4/test5")
remove.dir("~/Desktop/test2/test3/test4")
remove.dir("~/Desktop/test2/test3")
remove.dir("~/Desktop/test2")
remove.dir("~/Desktop/test")
}
\seealso{
Other directory management functions: 
\code{\link{remove.dir}()}
}
\concept{directory management functions}
