% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{file.remove_if_exists}
\alias{file.remove_if_exists}
\title{Delete if the File Exists}
\usage{
file.remove_if_exists(...)
}
\description{
Avoid throwing an error if a call to `file.remove()` is to a file does not exist by adding a conditional statement. A message is returned in the console for each file to confirm deletion or if the file was not found.
}
