% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionFileName.R
\name{FunctionFileName}
\alias{FunctionFileName}
\title{Function FileName}
\usage{
FunctionFileName()
}
\value{
dataframe with `fileName` and `functionName` columns
}
\description{
This function compares the name of the file in the R subdir within the working directory and with the name of the function within the file.
}
