% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_info.R
\name{get_kb_size}
\alias{get_kb_size}
\title{Get File Size in KB}
\usage{
get_kb_size(...)
}
\arguments{
\item{...}{character vectors containing file paths.  Tilde-expansion
    is done: see \code{\link[base]{path.expand}}.}
}
\value{
Dataframe of file path and size
}
\description{
Get File Size in KB
}
