% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{\%--\%}
\alias{\%--\%}
\title{Calculate distance between two points}
\usage{
x \%--\% ...
}
\arguments{
\item{a}{an object of class `point2d` or `point3d`}

\item{b}{an object of class `point2d` or `point3d`}
}
\description{
Calculate distance between two points
}
