% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{grepl_r_dir}
\alias{grepl_r_dir}
\title{Find R Directory Files that Match a Pattern}
\usage{
grepl_r_dir(
  recursive = FALSE,
  text_pattern,
  ignore.case = TRUE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)
}
\description{
Convenience wrapper around `grepl_text_files()` for an R directory within the current working directory from which the function is being called.
}
