% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{greplRScripts}
\alias{greplRScripts}
\title{Find R Scripts with a Pattern}
\usage{
greplRScripts(pattern)
}
\arguments{
\item{pattern}{pattern passed as the pattern argument in the grepl function}
}
\description{
Deprecated. This function identifies R scripts in the R subdirectory in the working directory that contains the pattern given as the argument.
}
