% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_mgmt.R
\name{list.files_only}
\alias{list.files_only}
\title{List Files without Subdirectories}
\usage{
list.files_only(dir, full.names = TRUE, recursive = TRUE)
}
\description{
List files in a given directory without including any folders.
}
