% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files_from.R
\name{list_files_from}
\alias{list_files_from}
\title{List Files From a Particular Date Based on Filename}
\usage{
list_files_from(date, ...)
}
\arguments{
\item{date}{Date in "YYYY-MM-DD" format}
}
\value{
Unlike the list.files function, this function returns a dataframe instead of a vector
}
\description{
List Files From a Particular Date Based on Filename
}
