% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cached_object.R
\name{load_cached_object}
\alias{load_cached_object}
\title{Load a cached object}
\usage{
load_cached_object(object_name)
}
\value{
cached object. NULL if no such cache exists. Console messages also accompany the output/
}
\description{
A cached object is retrieved based on a key generated by the path to the present R script, object name, and the working directory.
}
