% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFavicons.R
\name{makeFavicons}
\alias{makeFavicons}
\title{Build and Move Favicons}
\usage{
makeFavicons()
}
\description{
This package first runs the build_favicons() function in the pkgdown package, which auto-detects a logo.svg or logo.png file and runs it through an API. The function makes a copy of the favicon output to the "man/figures" directory for use.
}
