% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{move_text_to}
\alias{move_text_to}
\title{Append Text in One File to Another}
\usage{
move_text_to(to, ..., remove_files = FALSE, separator = "\\n\\n\\n\\n\\n")
}
\description{
Aggregate text from multiple files into an existing text file with the option to delete the originating file.
}
