% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{point}
\alias{point}
\title{Point class}
\usage{
point(x, y)
}
\arguments{
\item{x}{numeric, x coordinate}

\item{y}{numeric, y coordinate}
}
\value{
an object of class `point`
}
\description{
A S3 class for 2d spatial points
}
