% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{point2d}
\alias{point2d}
\title{2D Point class}
\usage{
point2d(x, y)
}
\arguments{
\item{x}{numeric, x coordinate}

\item{y}{numeric, y coordinate}
}
\value{
an object of class `point2d`
}
\description{
A S3 class for 2d spatial points
}
\examples{
library(tidyverse)
df <- data.frame(x = c(1, 2, 3, 4), y = c(4, 5, 2, 8))
mutate(df, p = map2(x, y, point2d))

}
