% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{point3d}
\alias{point3d}
\title{3D Point class}
\usage{
point3d(x, y, z)
}
\arguments{
\item{x}{numeric, x coordinate}

\item{y}{numeric, y coordinate}

\item{z}{numeric, z coordinate}
}
\value{
an object of class `point3d`
}
\description{
A S3 class for 3d spatial points
}
\examples{
library(tidyverse)
df <- data.frame(x = c(1, 2, 3, 4), y = c(4, 5, 2, 8), z = c(2, 5, 7, 8))
mutate(df, p = pmap(list(x, y, z), point3d))

}
