% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudioapi.R
\name{present_script_path}
\alias{present_script_path}
\title{Get the path to the present R script}
\usage{
present_script_path()
}
\description{
(Deprecated) This script returns the path of the R script in which it is being executed. It is helpful for documenting dataframes that have been written and README (see Readi package for implementation).
}
