% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_dir.R
\name{replace_dir}
\alias{replace_dir}
\title{Replace a folder}
\usage{
replace_dir(path_to_dir, path_to_dir_to_replace)
}
\arguments{
\item{path_to_dir}{path to the folder that is replacing}

\item{path_to_dir_to_replace}{path to folder that is being replaced}
}
\description{
Replace a folder
}
\details{
The name of the replacement folder will stay the same and will not take the name of path_to_dir if the names differ.
}
