% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{replace_ext}
\alias{replace_ext}
\title{Replace the Extension of a Filename}
\usage{
replace_ext(original_fn, replace_ext)
}
\arguments{
\item{original_fn}{Original filename including extension. For example, "IngestionFile.xlsx".}

\item{replace_ext}{Extension to replace the original filename's extension with. For example, "csv".}
}
\description{
Replace the Extension of a Filename
}
