% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-wd.R
\name{revert_wd}
\alias{revert_wd}
\title{Revert wd back to original wd}
\usage{
revert_wd(save = TRUE)
}
\description{
This function restores the original working directory when saved using the change_wd function.
}
