% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.R
\name{rm_all}
\alias{rm_all}
\title{Removes R objects in the global environment without warning if R object doesn't exist}
\usage{
rm_all(robj_names)
}
\arguments{
\item{robj_names}{vector of length 1 of greater of R object names}
}
\description{
Removes R objects in the global environment without warning if R object doesn't exist
}
