% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.R
\name{rm_all_end_with}
\alias{rm_all_end_with}
\title{Removes all objects in the global environment that have a certain pattern}
\usage{
rm_all_end_with(ends_with_pattern)
}
\arguments{
\item{ends_with_pattern}{pattern that the object name starts with}
}
\description{
Removes all objects in the global environment that have a certain pattern
}
