% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_all_start_with.R
\name{rm_all_objects_that_start_with}
\alias{rm_all_objects_that_start_with}
\title{Removes all objects in the global environment that have a certain pattern}
\usage{
rm_all_objects_that_start_with(starts_with_pattern)
}
\arguments{
\item{starts_with_pattern}{pattern that the object name starts with}
}
\description{
Removes all objects in the global environment that have a certain pattern
}
