% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_if_exists.R
\name{rm_if_exists}
\alias{rm_if_exists}
\title{Removes R objects in the global environment if R object doesn't exist}
\usage{
rm_if_exists(r_object, envir = globalenv())
}
\arguments{
\item{r_object}{single R object that is removed if it is present in the environment}
}
\description{
Removes R objects in the global environment if R object doesn't exist
}
