% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_file.R
\name{select_file}
\alias{select_file}
\title{Interactively select a file}
\usage{
select_file(dir, new_obj_name = NULL, full_path = FALSE)
}
\description{
Interactively select a file based on a menu printed into a console.
}
