% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_naming.R
\name{skin_file}
\alias{skin_file}
\title{Remove Path and Extension of a File}
\usage{
skin_file(file, rm_path = TRUE, rm_ext = TRUE)
}
\arguments{
\item{file}{String}

\item{rm_path}{Remove the path in the `file` argument?}

\item{rm_ext}{Remove the extension in the `file` argument?}
}
\description{
Remove Path and Extension of a File
}
