% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{string_to_vector}
\alias{string_to_vector}
\title{Parse and evaluate a string}
\usage{
string_to_vector(string)
}
\description{
(Deprecated) This is the counterpart function of vector_to_string in this package.
}
