% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_naming.R
\name{strip_fn}
\alias{strip_fn}
\title{Remove the file ext from fn}
\usage{
strip_fn(full_fn, rm_ext = TRUE, upto_nchar_file_ext = 4, rm_path = TRUE)
}
\arguments{
\item{full_fn}{filename}
}
\description{
Remove the file ext from fn
}
