% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{use_cache_key}
\alias{use_cache_key}
\title{Create the key for caching}
\usage{
use_cache_key(object_name)
}
\description{
The list object that serves as a key for the caching functions of the R.cache package. This function differs from cut_key in that this one is used to retrieve a cached object if it exists.
}
