% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_to_string.R
\name{vector_to_string}
\alias{vector_to_string}
\title{Convert a vector of values into a string}
\usage{
vector_to_string(vector_to_string)
}
\value{
a parseable string of length one
}
\description{
This function allows to store values neatly in an Excel or other dataframe for summary or record-keeping, but is in a format that can be readily parsed back to an expression if need be. This is important in cases where the names of columns want to be saved for a dataframe in a particular script to reference back or to QA later on. The string can be converted back to a vector by its counterpart function string_to_vector() in this package.
}
