% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{\%--\%}
\alias{\%--\%}
\title{Calculate distance between two points}
\usage{
x \%--\% ...
}
\arguments{
\item{a}{an object of class `point`}

\item{b}{an object of class `point`}
}
\description{
Calculate distance between two points
}
